+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ## Upcoming Changes (unreleased)
+- Nothing yet 
+
+## 0.5.0 - September 04, 2015
 - Added method ```valid_country``` for returning country for parsed phone just in case phone number was valid
 - Changed behavior of method ```country```, now it returns main country for international code in case there is such country, or first country from array
 - Added ```local_number``` method to ```Phone```, returns local number without area code
 - Added ```area_code``` method to ```Phone```, returns area code of phone or nil if none present for parsed number
-- Added ```extension``` method to ```Phone```, returns extension passed for parsing after "#" or ";" signs
+- Added ```extension``` method to ```Phone```, returns extension passed for parsing after ```#``` or ```;``` signs
+- Updated phones data
 
 ## 0.4.9 - July 28, 2015